Remove inflections, use active support
The form builder uses string inflections heavily to work its magic. After toying with the idea of rolling a simple version of string inflections to avoid having any external dependencies, I decided that this is simply not realistic in the case of string inflections, which are quite complex. Therefore, active support >= 4.0.0 will now be required for this library.